Improved cursor movement with language boxes
The cursor now tries to stay within the language it was prior
to the movement. If the cursor is inside a language box and is moved
to the end or the beginning of the language box it stays within the
box. If the cursor is outside of a box and is moved to the beginning
or end of a language box, it stays outside of the box. Only when crossing
the language border, does the cursor moved inside/outside a box.